# api/pkg/server/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a HelixML AI API.",
        "title": "HelixML API reference",
        "contact": {
            "name": "Helix support",
            "url": "https://app.tryhelix.ai/",
            "email": "info@helix.ml"
        },
        "version": "0.1",
        "x-logo": {
            "altText": "Helix logo",
            "url": "https://avatars.githubusercontent.com/u/149581110?s=200\u0026v=4"
        }
    },
    "host": "app.tryhelix.ai",
    "paths": {
...
        "/api/v1/sessions/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "description": "Request body with the message and model to start chat completion.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SessionChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sashabaranov_go-openai.FunctionCall": {
            "type": "object",
            "properties": {
                "arguments": {
                    "description": "call function with arguments in JSON format",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sashabaranov_go-openai.ToolCall": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/github_com_sashabaranov_go-openai.FunctionCall"
                },
                "id": {
                    "type": "string"
                },
                "index": {
                    "description": "Index is not nil only in chat completion chunk object",
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/github_com_sashabaranov_go-openai.ToolType"
                }
            }
        },
        "github_com_sashabaranov_go-openai.ToolType": {
            "type": "string",
            "enum": [
                "function"
            ],
            "x-enum-varnames": [
                "ToolTypeFunction"
            ]
        },
...
        "types.Choice": {
            "type": "object",
            "properties": {
                "delta": {
                    "$ref": "#/definitions/types.OpenAIMessage"
                },
                "finish_reason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "message": {
                    "$ref": "#/definitions/types.OpenAIMessage"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "types.CreatorType": {
            "type": "string",
            "enum": [
                "system",
                "assistant",
                "user",
                "tool"
            ],
            "x-enum-varnames": [
                "CreatorTypeSystem",
                "CreatorTypeAssistant",
                "CreatorTypeUser",
                "CreatorTypeTool"
            ]
        },
...
        "types.InteractionState": {
            "type": "string",
            "enum": [
                "",
                "waiting",
                "editing",
                "complete",
                "error"
            ],
            "x-enum-varnames": [
                "InteractionStateNone",
                "InteractionStateWaiting",
                "InteractionStateEditing",
                "InteractionStateComplete",
                "InteractionStateError"
            ]
        },
        "types.Item": {
            "type": "object",
            "properties": {
                "b64_json": {
                    "type": "string"
                },
                "embedding": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "index": {
                    "type": "integer"
                },
                "object": {
                    "type": "string"
                },
                "url": {
                    "description": "Images",
                    "type": "string"
                }
            }
        },
...
        "types.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "$ref": "#/definitions/types.MessageContent"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "Interaction ID",
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/types.CreatorType"
                },
                "state": {
                    "$ref": "#/definitions/types.InteractionState"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.MessageContent": {
            "type": "object",
            "properties": {
                "content_type": {
                    "description": "text, image, multimodal_text",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.MessageContentType"
                        }
                    ]
                },
                "parts": {
                    "description": "Parts is a list of strings or objects. For example for text, it's a list of strings, for\nmulti-modal it can be an object:\n\"parts\": [\n\t\t{\n\t\t\t\t\"content_type\": \"image_asset_pointer\",\n\t\t\t\t\"asset_pointer\": \"file-service://file-28uHss2LgJ8HUEEVAnXa70Tg\",\n\t\t\t\t\"size_bytes\": 185427,\n\t\t\t\t\"width\": 2048,\n\t\t\t\t\"height\": 1020,\n\t\t\t\t\"fovea\": null,\n\t\t\t\t\"metadata\": null\n\t\t},\n\t\t\"what is in the image?\"\n]",
                    "type": "array",
                    "items": {}
                }
            }
        },
        "types.MessageContentType": {
            "type": "string",
            "enum": [
                "text"
            ],
            "x-enum-varnames": [
                "MessageContentTypeText"
            ]
        },
        "types.OpenAIMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "The message content",
                    "type": "string"
                },
                "role": {
                    "description": "The message role",
                    "type": "string"
                },
                "tool_call_id": {
                    "description": "For Role=tool prompts this should be set to the ID given in the assistant's prior request to call a tool.",
                    "type": "string"
                },
                "tool_calls": {
                    "description": "For Role=assistant prompts this may be set to the tool calls generated by the model, such as function calls.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sashabaranov_go-openai.ToolCall"
                    }
                }
            }
        },
        "types.OpenAIResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Choice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Item"
                    }
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/types.OpenAIUsage"
                }
            }
        },
        "types.OpenAIUsage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
        "types.Provider": {
            "type": "string",
            "enum": [
                "openai",
                "togetherai",
                "helix",
                "vllm"
            ],
            "x-enum-varnames": [
                "ProviderOpenAI",
                "ProviderTogetherAI",
                "ProviderHelix",
                "ProviderVLLM"
            ]
        },
...
        "types.SessionChatRequest": {
            "type": "object",
            "properties": {
                "app_id": {
                    "description": "Assign the session settings from the specified app",
                    "type": "string"
                },
                "assistant_id": {
                    "description": "Which assistant are we speaking to?",
                    "type": "string"
                },
                "lora_dir": {
                    "type": "string"
                },
                "lora_id": {
                    "description": "the fine tuned data entity we produced from this session",
                    "type": "string"
                },
                "messages": {
                    "description": "Initial messages",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Message"
                    }
                },
                "model": {
                    "description": "The model to use",
                    "type": "string"
                },
                "organization_id": {
                    "description": "The organization this session belongs to, if any",
                    "type": "string"
                },
                "provider": {
                    "description": "The provider to use",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Provider"
                        }
                    ]
                },
                "rag_source_id": {
                    "type": "string"
                },
                "session_id": {
                    "description": "If empty, we will start a new session",
                    "type": "string"
                },
                "stream": {
                    "description": "If true, we will stream the response",
                    "type": "boolean"
                },
                "system": {
                    "description": "System message, only applicable when starting a new session",
                    "type": "string"
                },
                "tools": {
                    "description": "Available tools to use in the session",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "e.g. text, image",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.SessionType"
                        }
                    ]
                }
            }
        },
...
        "types.SessionType": {
            "type": "string",
            "enum": [
                "",
                "text",
                "image"
            ],
            "x-enum-varnames": [
                "SessionTypeNone",
                "SessionTypeText",
                "SessionTypeImage"
            ]
        },
...
    }
}